Initial Changes for confgurable ipsec.conf 	modified:   build.sh 	modified:   run.sh 	modified:   ../start.sh

+++ docker/run.sh
@@ -52,6 +52,7 @@ check_ip "$PRIVATE_IP" || PRIVATE_IP=$(ifconfig eth0 | grep -Eo 'inet (addr:)?([
 check_ip "$PRIVATE_IP" || exiterr "Cannot find valid private IP."
 
 # Create IPsec (Libreswan) config
+if [ ! -f cat > /etc/ipsec.conf ]; then
 cat > /etc/ipsec.conf <<EOF
 version 2.0
 
@@ -107,10 +108,15 @@ conn xauth-psk
   also=shared
 EOF
 
+fi
+
+
 # Specify IPsec PSK
+if [ ! -f cat > /etc/ipsec.secrets ]; then
 cat > /etc/ipsec.secrets <<EOF
 $PUBLIC_IP  %any  : PSK "$VPN_IPSEC_PSK"
 EOF
+fi
 
 # Create xl2tpd config
 cat > /etc/xl2tpd/xl2tpd.conf <<'EOF'
